Clean up 'conventions of commit messages' section.
This felt like it was a bit overly pedantic and specific without really
adding much of value here and seems like it might not encourage the best
commit messages. ("Add filename" isn't necessary a useful commit
message!)

I've replaced it with a few guidelines for good commit messages and an example.