Correct overly-aggressive default margins on figure elements.
Normalize.css introduces a default 40px left and right margin on figure
elements. This leads to overlapping issues on smaller screens, were the images
will break the page width and introduce horizontal scrolling. No good!

This fix attempts to provide a more sensible default behaviour to account for
different situations. The default figure element is given a left and right
margin of 0. Centre-aligned figures use the 'auto' margin value. Left- and
right-aligned captions inherit a little bit of margin (20px instead of 40)
in order to avoid butting up against text, but are otherwise aligned with
the edge of the text.

This provides a much better default and avoids any issues with horizontal
scrolling.